backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b3
azure-cognitiveservices-speech==1.30.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.34
langchain-text-splitters==0.2.2
langsmith==0.1.101
MarkupSafe==2.1.5
marshmallow==3.22.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.41.1
orjson==3.10.7
packaging==24.1
portalocker==2.10.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
